+++ lib/enola-guides.rb
@@ -2,6 +2,6 @@
 
 module Enola
   module Guides
-    VERSION = "0.1.1"
+    VERSION = "0.1.2"
   end
 end